+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## [Unreleased][]
 ### Added
+- UI: added background framework found in all my other scripts, which can be tweaked at the background submenu within the toolbar menu (R. Click). By default is set to to not display any color or art, looking like previous releases. Note any setting at 'Colors' submenu is applied on top of the background, so the toolbar color, transparency or panel pseudo-transparency settings may also override background drawing. As a bonus, all features related to dynamic colors and color servers are also included, so it will work in conjunction with all other scripts making use of this feature.
 ### Changed
 ### Removed
 ### Fixed